assignment2_2023/assignment2/requirements.txt
haversine==2.8.0
mysql-connector-python==8.1.0
tabulate==0.9.0
protobuf==4.21.12